pyproject.toml
[tool.poetry]
name = "payroll-app"
version = "0.1.0"
description = ""
authors = ["Semyon <odintsovvans@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
pydantic = "^2.9.2"
pytest = "^8.3.3"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-storages = "^0.3.0"
greenlet = "^3.1.1"
python-multipart = "^0.0.12"
python-jose = "^3.3.0"
fastapi-jwt = "^0.3.0"
pytest-asyncio = "^0.24.0"
aiosmtplib = "^3.0.2"
aioredis = "^2.0.1"
redis = "^5.2.0"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
pythonpath = "src/"
addopts = "-vv -p no:cacheprovider --disable-warnings"
testpaths = ["tests/"]
python_files = ["test_*.py"]
asyncio_mode= "auto"
